requirements.txt
Flask>=3.0.0
psutil>=5.9.0
netifaces>=0.11.0
speedtest-cli>=2.1.3
requests>=2.31.0
